Move MockCallInterfaceConnector
